requirements.txt
streamlit
langchain-core
langchain-community
langchain-cohere
cohere
PyPDF2
python-dotenv
faiss-cpu
langchain-groq
sentence-transformers
tiktoken